requirements.txt
amqp==5.1.1
APScheduler==3.6.3
async-timeout==4.0.2
billiard==3.6.4.0
cachetools==4.2.2
celery==5.2.7
certifi==2022.6.15
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect==1.1.0
Deprecated==1.2.13
feedparser==6.0.10
kombu==5.2.4
lxml==4.9.1
packaging==21.3
prompt-toolkit==3.0.30
pyparsing==3.0.9
pyquery==1.4.3
python-dotenv==0.20.0
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
sgmllib3k==1.0.0
six==1.16.0
tornado==6.2
tzdata==2022.1
tzlocal==4.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
